make agent its own class file
